Added Initial Ribbon Elements
Added two groups boxes, one for Reports and one for Estimates.  They contain one and two buttons, respectively.  For Report options, one button opens a SaveFileDialog and the other opens a FileDialog filtered for CSV files.  The FileDialog is the key feature of today's changes, since it retrieves the filepath to the field report.  The file path is handed over to the CSVDataService utility, which reads its contents into the ProgressReporter methods that will produce the report output.

Mike is currently working on the CSVDataService using the information we covered in yesterday's meeting.